refactor: simplified models definition further
