+++ py_schemax/validator.py
@@ -9,8 +9,11 @@ from pydantic import ValidationError
 from pydantic_core import ErrorDetails
 
 from py_schemax.config import Config
-from py_schemax.model import SupportedDataTypes, get_dynamic_dataset_schema
-from py_schemax.models import DatasetSchema
+from py_schemax.model import (
+    DatasetSchema,
+    SupportedDataTypes,
+    get_dynamic_dataset_schema,
+)
 
 
 class PydanticErrorSchema(TypedDict):
